// src/interfaces/pokemon/pokemon-info.ts
export interface PokemonInfo {
  attack?: string;
  candytype?: string;
  capture?: string;
  cp?: string;
  defense?: string;
  flee?: string;
  height?: string;
  image?: string;
  needevolve?: string;
  nrmheight?: string;
  nrmweight?: string;
  pokedex?: string;
  stage1?: string;
  stage2?: string;
  stage3?: string;
  stamina?: string;
  type1?: string;
  type2?: string;
  weak?: string;
  weight?: string;
  xlheight?: string;
  xlweight?: string;
  xsweight?: string;
}